Update maven-compiler-plugin
Also, added commented-out module-info and commented out javafx 11 jars to pom.xml. Should be a good start to Java upgrade but was unsuccessful, there is more to it than just this. There may also be issues with certain dependencies not yet in modular format... eventually upgrading the project may be something to do but not today.